requirement.txt
pillow
pyzbar
opencv-python
pandas
glob2
keyboard
openpyx
time